requirements.txt
django>=3.1
django-crispy-forms>=1.9
django-q>=1.3.3
torch
nltk
scikit-learn
pytorch-pretrained-bert